外部サービスに関連するモデルは external_service という名前空間を作って、そこに属するようにクラス構造を変更した
