+++ app/models/external_service/base.rb
@@ -1,0 +1,5 @@
+module ExternalService
+  class Base
+    SEARCH_TRACKS_NUMBER = 5
+  end
+end
