+++ app/models/response.rb
@@ -1,0 +1,7 @@
+class Response
+  attr_reader :body
+
+  def initialize(response)
+    @body = JSON.parse(response)
+  end
+end
